DRY Provisioner
